+++ src/core/interface/index.ts
@@ -1,9 +1,0 @@
-export interface BaseField {
-  id?: string;
-  field: string;
-  prop?: string;
-  show?: boolean;
-  slotData?: Record<string, any>;
-  children?: (BaseField | any)[];
-  slots?: Record<string, (BaseField | any)[]>;
-}